+++ Layouts/BlizzCompact.lua
@@ -212,7 +212,7 @@ function layout:Initialize(frame)
 
     if not classIcon.ClassIconBorderHook then
         hooksecurefunc(classIcon, "SetTexture", function(self, t)
-            if t == nil or t == "" or t == 0 or t == "nil" or not self.useModernBorder then
+            if not t or not self.useModernBorder then
                 classIconBorder:Hide()
             else
                 classIconBorder:Hide()
@@ -252,7 +252,7 @@ function layout:Initialize(frame)
 
     if not trinket.TrinketBorderHook then
         hooksecurefunc(trinket.Texture, "SetTexture", function(self, t)
-            if t == nil or t == "" or t == 0 or t == "nil" or not trinket.useModernBorder then
+            if not t or not trinket.useModernBorder then
                 trinketBorder:Hide()
             else
                 trinketBorder:Hide()
@@ -293,7 +293,7 @@ function layout:Initialize(frame)
 
     if not racial.RacialBorderHook then
         hooksecurefunc(racial.Texture, "SetTexture", function(self, t)
-            if t == nil or t == "" or t == 0 or t == "nil" or not racial.useModernBorder then
+            if not t or not racial.useModernBorder then
                 racialBorder:Hide()
             else
                 racialBorder:Hide()
@@ -336,7 +336,7 @@ function layout:Initialize(frame)
 
     if not dispel.DispelBorderHook then
         hooksecurefunc(dispel.Texture, "SetTexture", function(self, t)
-            if t == nil or t == "" or t == 0 or t == "nil" or not dispel.useModernBorder then
+            if not t or not dispel.useModernBorder then
                 dispelBorder:Hide()
             else
                 dispelBorder:Hide()